bump version of ExtUtils::ParseXS
This bumps the version of ExtUtils::ParseXS to the first refactored
stable version of the module (ExtUtils::ParseXS v3.01). The build was
previously failing on the version that came with Perl v5.12
(ExtUtils::ParseXS v2.21).

Though it might be easier to try to fix the build itself so that the
older ExtUtils::ParseXS is able to work, explicitly using a newer
version may help avoid other build errors.